End of the day push. I at least have the app appearing on my screen. Need to fix attendees.js. Cosider also moving some of the functionality into the other components.
