+++ src/App.js
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import eventsData from "./data";
 import { v1 as generateUniqueID } from "uuid";
-import Attendees from "./Attendees";
+import Attendees from "./Components/Attendees";
 import Event from "./Components/Event";
 import Footer from "./Components/Footer";
 import Header from "./Components/Header";
@@ -37,6 +37,8 @@ function App() {
     handleAddEvent(createEvent);
   }
 
+
+  //NOTE: why is this never used?
   function handleSelectChange(e) {
     setSelectOption(e.target.value);
   }
